+++ src/app_service.rs
@@ -21,8 +21,6 @@ use crate::{
     Error, HttpResponse,
 };
 
-type Guards = Vec<Box<dyn Guard>>;
-
 /// Service factory to convert `Request` to a `ServiceRequest<S>`.
 ///
 /// It also executes data factories.
@@ -244,7 +242,7 @@ pub struct AppRoutingFactory {
         [(
             ResourceDef,
             BoxedHttpServiceFactory,
-            RefCell<Option<Guards>>,
+            RefCell<Option<Vec<Box<dyn Guard>>>>,
         )],
     >,
     default: Rc<BoxedHttpServiceFactory>,
@@ -262,7 +260,7 @@ impl ServiceFactory<ServiceRequest> for AppRoutingFactory {
         // construct all services factory future with it's resource def and guards.
         let factory_fut = join_all(self.services.iter().map(|(path, factory, guards)| {
             let path = path.clone();
-            let guards = guards.borrow_mut().take();
+            let guards = guards.borrow_mut().take().unwrap_or_default();
             let factory_fut = factory.new_service(());
             async move {
                 let service = factory_fut.await?;
@@ -295,7 +293,7 @@ impl ServiceFactory<ServiceRequest> for AppRoutingFactory {
 
 /// The Actix Web router default entry point.
 pub struct AppRouting {
-    router: Router<BoxedHttpService, Option<Guards>>,
+    router: Router<BoxedHttpService, Vec<Box<dyn Guard>>>,
     default: BoxedHttpService,
 }
 
@@ -309,7 +307,7 @@ impl Service<ServiceRequest> for AppRouting {
     fn call(&self, mut req: ServiceRequest) -> Self::Future {
         let res = self.router.recognize_fn(&mut req, |req, guards| {
             let guard_ctx = req.guard_ctx();
-            guards.iter().flatten().all(|guard| guard.check(&guard_ctx))
+            guards.iter().all(|guard| guard.check(&guard_ctx))
         });
 
         if let Some((srv, _info)) = res {